[DEV] update Slow GPX

+++ pygpxtools/pygpxtools.py
@@ -186,8 +186,9 @@ def cli_change_timestamps(input, output, year, month, day, hour, minute, second)
 @click.option('--input', help='Input GPX file from Garmin Connect where remove pauses', default=None)
 @click.option('--output', help='Output GPX file to upload to Strava',
               default='/home/alexantr/tmp/pygpxtools_' + datetime.datetime.today().strftime('%Y%m%d%H%M') + '.gpx')
-@click.option('--factor', help='Slow down inter point factor in milliseconds', default=100)
-def cli_slow(input, output, factor):
+@click.option('--factor', help='Slow down factor', type=float)
+@click.option('--dryrun', help='do not generate output', is_flag=True)
+def cli_slow(input, output, factor, dryrun):
     """
     Decrease speed in Garmin GPX file in adding factor to current timestamp.
     Only support files with .gpx extension.
@@ -203,16 +204,28 @@ def cli_slow(input, output, factor):
     input = check_input_file(input, 'slowDown')
     output = check_output_file(output)
 
-    correction = factor
     with open(input, 'r') as gpx_file:
         gpx = gpxpy.parse(gpx_file)
+        points_no = len(list(gpx.walk(only_points=True)))
+        moving_time, stopped_time, moving_distance, stopped_distance, max_speed = gpx.get_moving_data()
+        print('Debug#01: numbers of point: {}'.format(points_no))
+        print('Debug#02: total time: {}'.format(format_time(moving_time)))
+        correction = step = ((moving_time * factor / 100) / (points_no - 1)) * 1000  # convert to milliseconds
+        print('Debug#03: factor {} - correction of {}s between 2 points.'.format(factor, step))
         for track in gpx.tracks:
             for segment in track.segments:
+                is_first_point = True
                 for point in segment.points:
-                    point.time = point.time + datetime.timedelta(milliseconds=correction)
-                    correction += factor
-        with open(output, 'w') as new_file:
-            new_file.write(gpx.to_xml())
+                    if not is_first_point:
+                        # print('Debug#04: current point time {} - new point time {} - correction: {}'.format(point.time, point.time + datetime.timedelta(milliseconds=correction), correction))
+                        point.time = point.time + datetime.timedelta(milliseconds=correction)
+                        correction += step
+                    else:
+                        # print('Debug#04: 1st Point : current point time {}'.format(point.time))
+                        is_first_point = False
+        if not dryrun:
+            with open(output, 'w') as new_file:
+                new_file.write(gpx.to_xml())
 
 
 @cli.command('stravaUpload')
@@ -241,7 +254,7 @@ def cli_strava_upload(input, login, password):
             if gpx.time.hour < 6:
                 activity_title += 'la nuit'
             elif 6 <= gpx.time.hour < 12:
-                activity_title += 'le matin'
+                activity_title += 'matinale'
             elif 12 <= gpx.time.hour < 14:
                 activity_title += 'le midi'
             elif 14 <= gpx.time.hour < 19:
